+++ test/TwigRendererFactoryTest.php
@@ -1,7 +1,7 @@
 <?php
 /**
  * @see       https://github.com/zendframework/zend-expressive-twigrenderer for the canonical source repository
- * @copyright Copyright (c) 2015-2017 Zend Technologies USA Inc. (https://www.zend.com)
+ * @copyright Copyright (c) 2015-2018 Zend Technologies USA Inc. (https://www.zend.com)
  * @license   https://github.com/zendframework/zend-expressive-twigrenderer/blob/master/LICENSE.md New BSD License
  */
 
@@ -17,7 +17,9 @@ use Twig_Environment as TwigEnvironment;
 use Zend\Expressive\Helper\ServerUrlHelper;
 use Zend\Expressive\Helper\UrlHelper;
 use Zend\Expressive\Template\TemplatePath;
+use Zend\Expressive\Twig\Exception\InvalidConfigException;
 use Zend\Expressive\Twig\TwigEnvironmentFactory;
+use Zend\Expressive\Twig\TwigExtension;
 use Zend\Expressive\Twig\TwigRenderer;
 use Zend\Expressive\Twig\TwigRendererFactory;
 
@@ -122,6 +124,7 @@ class TwigRendererFactoryTest extends TestCase
     public function testCallingFactoryWithNoConfigReturnsTwigInstance()
     {
         $this->container->has('config')->willReturn(false);
+        $this->container->has(TwigExtension::class)->willReturn(false);
         $this->container->has(ServerUrlHelper::class)->willReturn(false);
         $this->container->has(UrlHelper::class)->willReturn(false);
         $environment = new TwigEnvironmentFactory();
@@ -158,6 +161,7 @@ class TwigRendererFactoryTest extends TestCase
         ];
         $this->container->has('config')->willReturn(true);
         $this->container->get('config')->willReturn($config);
+        $this->container->has(TwigExtension::class)->willReturn(false);
         $this->container->has(ServerUrlHelper::class)->willReturn(false);
         $this->container->has(UrlHelper::class)->willReturn(false);
         $environment = new TwigEnvironmentFactory();
@@ -180,6 +184,7 @@ class TwigRendererFactoryTest extends TestCase
         ];
         $this->container->has('config')->willReturn(true);
         $this->container->get('config')->willReturn($config);
+        $this->container->has(TwigExtension::class)->willReturn(false);
         $this->container->has(ServerUrlHelper::class)->willReturn(false);
         $this->container->has(UrlHelper::class)->willReturn(false);
         $environment = new TwigEnvironmentFactory();
@@ -210,6 +215,7 @@ class TwigRendererFactoryTest extends TestCase
     public function testCallingFactoryWithoutTwigEnvironmentServiceEmitsDeprecationNotice()
     {
         $this->container->has('config')->willReturn(false);
+        $this->container->has(TwigExtension::class)->willReturn(false);
         $this->container->has(ServerUrlHelper::class)->willReturn(false);
         $this->container->has(UrlHelper::class)->willReturn(false);
         $this->container->has(TwigEnvironment::class)->willReturn(false);
@@ -223,5 +229,42 @@ class TwigRendererFactoryTest extends TestCase
 
         $twig = $factory($this->container->reveal());
         $this->assertInstanceOf(TwigRenderer::class, $twig);
+    }
+
+    public function testMergeConfigRaisesExceptionForInvalidConfig()
+    {
+        $this->expectException(InvalidConfigException::class);
+        $this->expectExceptionMessage('Config service MUST be an array or ArrayObject; received string');
+
+        TwigRendererFactory::mergeConfig('foo');
+    }
+
+    public function testMergesConfigCorrectly()
+    {
+        $config = [
+            'templates' => [
+                'extension' => 'file extension used by templates; defaults to html.twig',
+                'paths' => [],
+            ],
+            'twig' => [
+                'cache_dir' => 'path to cached templates',
+                'assets_url' => 'base URL for assets',
+                'assets_version' => 'base version for assets',
+                'extensions' => [],
+                'runtime_loaders' => [],
+                'globals' => ['ga_tracking' => 'UA-XXXXX-X'],
+                'timezone' => 'default timezone identifier, e.g.: America/New_York',
+            ],
+        ];
+
+        $mergedConfig = TwigRendererFactory::mergeConfig($config);
+
+        $this->assertArrayHasKey('extension', $mergedConfig);
+        $this->assertArrayHasKey('paths', $mergedConfig);
+        $this->assertArrayHasKey('cache_dir', $mergedConfig);
+        $this->assertArrayHasKey('assets_version', $mergedConfig);
+        $this->assertArrayHasKey('runtime_loaders', $mergedConfig);
+        $this->assertArrayHasKey('globals', $mergedConfig);
+        $this->assertArrayHasKey('timezone', $mergedConfig);
     }
 }